# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JobProfile {
  id: number;
  refNo: string;
  title: string;
  department?: string;
  description: string;
  requiredSkills: string[];
  preferredSkills: string[];
  minExperienceYears: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateJobProfileBody {
  refNo: string;
  title: string;
  department?: string;
  description: string;
  requiredSkills: string[];
  preferredSkills: string[];
  minExperienceYears: number;
}

export interface Resume {
  id: number;
  refNo: string;
  fileName: string;
  originalText: string;
  candidateName?: string;
  candidateEmail?: string;
  candidatePhone?: string;
  candidateAddress?: string;
  uploadedAt: string;
  analysisId?: number | null;
}

export interface AnalysisResult {
  id: number;
  resumeId: number;
  refNo: string;
  candidateName: string;
  candidateEmail?: string;
  candidatePhone?: string;
  candidateAddress?: string;
  profileSummary: string;
  atsScore: number;
  suitability: string;
  matchingSkills: string[];
  skillsGap: string[];
  experienceSummary: string;
  yearsExperience: number;
  recommendation: string;
  analyzedAt: string;
}

export interface BatchAnalysisBody {
  refNo: string;
}

export interface SummaryReport {
  refNo: string;
  jobTitle: string;
  generatedAt: string;
  totalCandidates: number;
  candidates: AnalysisResult[];
}

export type ListResumesParams = {
  refNo?: string;
};

export type UploadResumeBody = {
  file: Blob;
  refNo: string;
  fileName?: string;
};