# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all job profiles
 */
export const ListJobProfilesResponseItem = zod.object({
  id: zod.number(),
  refNo: zod.string(),
  title: zod.string(),
  department: zod.string().optional(),
  description: zod.string(),
  requiredSkills: zod.array(zod.string()),
  preferredSkills: zod.array(zod.string()),
  minExperienceYears: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListJobProfilesResponse = zod.array(ListJobProfilesResponseItem);

/**
 * @summary Create a job profile
 */
export const CreateJobProfileBody = zod.object({
  refNo: zod.string(),
  title: zod.string(),
  department: zod.string().optional(),
  description: zod.string(),
  requiredSkills: zod.array(zod.string()),
  preferredSkills: zod.array(zod.string()),
  minExperienceYears: zod.number(),
});

/**
 * @summary Get job profile by reference number
 */
export const GetJobProfileParams = zod.object({
  refNo: zod.coerce.string(),
});

export const GetJobProfileResponse = zod.object({
  id: zod.number(),
  refNo: zod.string(),
  title: zod.string(),
  department: zod.string().optional(),
  description: zod.string(),
  requiredSkills: zod.array(zod.string()),
  preferredSkills: zod.array(zod.string()),
  minExperienceYears: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Update a job profile
 */
export const UpdateJobProfileParams = zod.object({
  refNo: zod.coerce.string(),
});

export const UpdateJobProfileBody = zod.object({
  refNo: zod.string(),
  title: zod.string(),
  department: zod.string().optional(),
  description: zod.string(),
  requiredSkills: zod.array(zod.string()),
  preferredSkills: zod.array(zod.string()),
  minExperienceYears: zod.number(),
});

export const UpdateJobProfileResponse = zod.object({
  id: zod.number(),
  refNo: zod.string(),
  title: zod.string(),
  department: zod.string().optional(),
  description: zod.string(),
  requiredSkills: zod.array(zod.string()),
  preferredSkills: zod.array(zod.string()),
  minExperienceYears: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a job profile
 */
export const DeleteJobProfileParams = zod.object({
  refNo: zod.coerce.string(),
});

/**
 * @summary List all resumes
 */
export const ListResumesQueryParams = zod.object({
  refNo: zod.coerce.string().optional(),
});

export const ListResumesResponseItem = zod.object({
  id: zod.number(),
  refNo: zod.string(),
  fileName: zod.string(),
  originalText: zod.string(),
  candidateName: zod.string().optional(),
  candidateEmail: zod.string().optional(),
  candidatePhone: zod.string().optional(),
  candidateAddress: zod.string().optional(),
  uploadedAt: zod.date(),
  analysisId: zod.number().nullish(),
});
export const ListResumesResponse = zod.array(ListResumesResponseItem);

/**
 * @summary Upload a resume file
 */
export const UploadResumeBody = zod.object({
  file: zod.instanceof(File),
  refNo: zod.string(),
  fileName: zod.string().optional(),
});

/**
 * @summary Get a resume by ID
 */
export const GetResumeParams = zod.object({
  id: zod.coerce.number(),
});

export const GetResumeResponse = zod.object({
  id: zod.number(),
  refNo: zod.string(),
  fileName: zod.string(),
  originalText: zod.string(),
  candidateName: zod.string().optional(),
  candidateEmail: zod.string().optional(),
  candidatePhone: zod.string().optional(),
  candidateAddress: zod.string().optional(),
  uploadedAt: zod.date(),
  analysisId: zod.number().nullish(),
});

/**
 * @summary Delete a resume
 */
export const DeleteResumeParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Run ATS analysis on a resume
 */
export const AnalyzeResumeParams = zod.object({
  resumeId: zod.coerce.number(),
});

export const AnalyzeResumeResponse = zod.object({
  id: zod.number(),
  resumeId: zod.number(),
  refNo: zod.string(),
  candidateName: zod.string(),
  candidateEmail: zod.string().optional(),
  candidatePhone: zod.string().optional(),
  candidateAddress: zod.string().optional(),
  profileSummary: zod.string(),
  atsScore: zod.number(),
  suitability: zod.string(),
  matchingSkills: zod.array(zod.string()),
  skillsGap: zod.array(zod.string()),
  experienceSummary: zod.string(),
  yearsExperience: zod.number(),
  recommendation: zod.string(),
  analyzedAt: zod.date(),
});

/**
 * @summary Run ATS analysis on all resumes for a reference number
 */
export const AnalyzeBatchBody = zod.object({
  refNo: zod.string(),
});

export const AnalyzeBatchResponseItem = zod.object({
  id: zod.number(),
  resumeId: zod.number(),
  refNo: zod.string(),
  candidateName: zod.string(),
  candidateEmail: zod.string().optional(),
  candidatePhone: zod.string().optional(),
  candidateAddress: zod.string().optional(),
  profileSummary: zod.string(),
  atsScore: zod.number(),
  suitability: zod.string(),
  matchingSkills: zod.array(zod.string()),
  skillsGap: zod.array(zod.string()),
  experienceSummary: zod.string(),
  yearsExperience: zod.number(),
  recommendation: zod.string(),
  analyzedAt: zod.date(),
});
export const AnalyzeBatchResponse = zod.array(AnalyzeBatchResponseItem);

/**
 * @summary Get summary report for a reference number
 */
export const GetSummaryReportParams = zod.object({
  refNo: zod.coerce.string(),
});

export const GetSummaryReportResponse = zod.object({
  refNo: zod.string(),
  jobTitle: zod.string(),
  generatedAt: zod.date(),
  totalCandidates: zod.number(),
  candidates: zod.array(
    zod.object({
      id: zod.number(),
      resumeId: zod.number(),
      refNo: zod.string(),
      candidateName: zod.string(),
      candidateEmail: zod.string().optional(),
      candidatePhone: zod.string().optional(),
      candidateAddress: zod.string().optional(),
      profileSummary: zod.string(),
      atsScore: zod.number(),
      suitability: zod.string(),
      matchingSkills: zod.array(zod.string()),
      skillsGap: zod.array(zod.string()),
      experienceSummary: zod.string(),
      yearsExperience: zod.number(),
      recommendation: zod.string(),
      analyzedAt: zod.date(),
    }),
  ),
});

/**
 * @summary Export summary report as CSV
 */
export const ExportReportParams = zod.object({
  refNo: zod.coerce.string(),
});