# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisResult {
  id: number;
  resumeId: number;
  refNo: string;
  candidateName: string;
  candidateEmail?: string;
  candidatePhone?: string;
  candidateAddress?: string;
  profileSummary: string;
  atsScore: number;
  suitability: string;
  matchingSkills: string[];
  skillsGap: string[];
  experienceSummary: string;
  yearsExperience: number;
  recommendation: string;
  analyzedAt: Date;
}