# lib/api-zod/src/generated/types/createJobProfileBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJobProfileBody {
  refNo: string;
  title: string;
  department?: string;
  description: string;
  requiredSkills: string[];
  preferredSkills: string[];
  minExperienceYears: number;
}