# lib/api-zod/src/generated/types/resume.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */

export interface Resume {
  id: number;
  refNo: string;
  fileName: string;
  originalText: string;
  candidateName?: string;
  candidateEmail?: string;
  candidatePhone?: string;
  candidateAddress?: string;
  uploadedAt: Date;
  analysisId?: number | null;
}