# lib/api-zod/src/generated/types/summaryReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ATS Resume Checker API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from "./analysisResult";

export interface SummaryReport {
  refNo: string;
  jobTitle: string;
  generatedAt: Date;
  totalCandidates: number;
  candidates: AnalysisResult[];
}